Restructure PurchaseContentForm to be sectionable by ModalContent and ModalFooter
