+++ packages/web/src/components/premium-content-purchase-modal/components/PayExtraFormSection.tsx
@@ -40,6 +40,7 @@ export const PayExtraFormSection = ({
             isSelected={preset === PayExtraPreset.LOW}
             label={formatPillAmount(amountPresets[PayExtraPreset.LOW])}
             size='large'
+            type='button'
             onClick={() => handleClickPreset(PayExtraPreset.LOW)}
           />
           <HarmonySelectablePill
@@ -47,6 +48,7 @@ export const PayExtraFormSection = ({
             isSelected={preset === PayExtraPreset.MEDIUM}
             label={formatPillAmount(amountPresets[PayExtraPreset.MEDIUM])}
             size='large'
+            type='button'
             onClick={() => handleClickPreset(PayExtraPreset.MEDIUM)}
           />
           <HarmonySelectablePill
@@ -54,6 +56,7 @@ export const PayExtraFormSection = ({
             isSelected={preset === PayExtraPreset.HIGH}
             label={formatPillAmount(amountPresets[PayExtraPreset.HIGH])}
             size='large'
+            type='button'
             onClick={() => handleClickPreset(PayExtraPreset.HIGH)}
           />
         </div>
@@ -62,6 +65,7 @@ export const PayExtraFormSection = ({
           isSelected={preset === PayExtraPreset.CUSTOM}
           label={messages.customAmount}
           size='large'
+          type='button'
           onClick={() => handleClickPreset(PayExtraPreset.CUSTOM)}
         />
       </div>
