--- packages/web/src/components/premium-content-purchase-modal/components/PurchaseContentForm.module.css
@@ -1,44 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: var(--unit-6);
-}
-
-.purchaseButtonText {
-  align-items: center;
-  display: inline-flex;
-  gap: var(--unit-2);
-}
-
-.purchaseButtonSpinner {
-  width: var(--unit-5);
-  height: var(--unit-5);
-}
-
-.purchaseButtonSpinner g path {
-  stroke: currentColor;
-}
-
-.purchaseSuccessfulContainer {
-  align-items: center;
-  display: flex;
-  gap: var(--unit-2);
-  justify-content: center;
-  padding: var(--unit-3);
-}
-
-.completionCheck {
-  align-items: center;
-  background: var(--special-green);
-  border-radius: 50%;
-  display: flex;
-  height: var(--unit-4);
-  justify-content: center;
-  width: var(--unit-4);
-}
-
-.errorContainer {
-  display: flex;
-  gap: var(--unit-2);
-  align-items: center;
-}
