+++ packages/web/src/components/premium-content-purchase-modal/components/PurchaseContentFormFields.module.css
@@ -1,0 +1,17 @@
+.purchaseSuccessfulContainer {
+  align-items: center;
+  display: flex;
+  gap: var(--unit-2);
+  justify-content: center;
+  padding: var(--unit-3);
+}
+
+.completionCheck {
+  align-items: center;
+  background: var(--special-green);
+  border-radius: 50%;
+  display: flex;
+  height: var(--unit-4);
+  justify-content: center;
+  width: var(--unit-4);
+}
