+++ packages/web/src/components/premium-content-purchase-modal/components/PurchaseContentFormFooter.module.css
@@ -1,0 +1,20 @@
+.purchaseButtonText {
+  align-items: center;
+  display: inline-flex;
+  gap: var(--unit-2);
+}
+
+.spinner {
+  width: var(--unit-5);
+  height: var(--unit-5);
+}
+
+.spinner g path {
+  stroke: currentColor;
+}
+
+.errorContainer {
+  display: flex;
+  gap: var(--unit-2);
+  align-items: center;
+}
